WIP: Input states
- In progress work in compose script input feature
- Top level language construct that allows entities and components to
  define input states that affect them.
- Should simplify input state management of scripts and give a really
  nice concise way of hooking game states to inputs.

+++ src/engine/input.cpp
@@ -192,8 +192,8 @@ void kill_focus()
 
 static const f32 kMouseUpdateInterval = 1.0f / 60.0f;
 static f32 sLastUpdateTime = 0.0f;
-static short sMouseDeltaX;
-static short sMouseDeltaY;
+static i32 sMouseDeltaX;
+static i32 sMouseDeltaY;
 
 void send_mouse_input()
 {
